feature. https://gitlab.iptco.me/tio/wip/-/issues/174 收藏列表

+++ module_hall/src/main/java/com/walisport/module/hall/ui/fragment/GameAllRankingTodayFragment.kt
@@ -56,7 +56,7 @@ class GameAllRankingTodayFragment :
 
             //如果是华为手机，增加顶部间距，避免被刘海遮挡
             val params= rvTodayRank.layoutParams as ConstraintLayout.LayoutParams
-            params.topMargin = (-10f).dp2px
+            params.topMargin = (-10f).dp2px.toInt()
             rvTodayRank.layoutParams = params
         }
     }
